[Tests] Separate mock classes out
