+++ tests/phpunit/unit/Events/Mock/ClippyKoala.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Bolt\Tests\Events\Mock;
+
+/**
+ * Clippy vs. Koala … The final showdown!
+ *
+ * @author Gawain Lynch <gawain.lynch@gmail.com>
+ */
+class ClippyKoala
+{
+}
